+++ crates/storage/codecs/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 [dependencies]
 # reth
 reth-codecs-derive = { path = "./derive", default-features = false }
+reth-db-serialization.workspace = true
 
 # eth
 alloy-consensus = { workspace = true, optional = true }
@@ -28,7 +29,6 @@ op-alloy-consensus = { workspace = true, optional = true }
 bytes.workspace = true
 modular-bitfield = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
-derive_more.workspace = true
 
 [dev-dependencies]
 alloy-eips = { workspace = true, default-features = false, features = [
