+++ crates/storage/db-serialization/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "reth-db-serialization"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# eth
+alloy-primitives.workspace = true
+
+#misc
+derive_more.workspace = true
+
+[features]
+default = ["std"]
+std = ["alloy-primitives/std"]
