+++ crates/storage/db-serialization/src/encode.rs
